backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
ib_insync==0.9.86
nest_asyncio==1.6.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
pytz==2024.1
# Production dependencies
asyncpg==0.30.0
PyJWT==2.10.1
httpx==0.28.1
cryptography==44.0.0
websockets==14.1
